Improve some remaining side-effects and update tests for telling us which modules have side-effects
