+++ packages/@ember/-internals/glimmer/lib/helper.ts
@@ -10,12 +10,14 @@ import getDebugName from '@ember/-internals/utils/lib/get-debug-name';
 import { assert } from '@ember/debug';
 import { join } from '@ember/runloop';
 import type { Arguments, HelperManager } from '@glimmer/interfaces';
-import { getInternalHelperManager } from '@glimmer/manager/lib/internal/api';
 import { helperCapabilities } from '@glimmer/manager/lib/public/helper';
 import { setHelperManager } from '@glimmer/manager/lib/public/api';
 import type { DirtyableTag } from '@glimmer/interfaces';
 import { consumeTag } from '@glimmer/validator/lib/tracking';
 import { createTag, DIRTY_TAG as dirtyTag } from '@glimmer/validator/lib/validators';
+import { IS_CLASSIC_HELPER } from './helper-brand';
+
+export { isClassicHelper } from './helper-brand';
 
 const RECOMPUTE_TAG = Symbol('RECOMPUTE_TAG');
 
@@ -44,8 +46,6 @@ export interface HelperInstance<S> {
   [RECOMPUTE_TAG]: DirtyableTag;
 }
 
-const IS_CLASSIC_HELPER: unique symbol = Symbol('IS_CLASSIC_HELPER');
-
 export interface SimpleHelper<S> {
   compute: (positional: Positional<S>, named: Named<S>) => Return<S>;
 }
@@ -60,7 +60,7 @@ declare const SIGNATURE: unique symbol;
 
   ```app/templates/application.gjs
   import Cost from '../components/cost';
-    
+
   <template>
     <Cost @cents={{230}} />
   </template>
@@ -68,7 +68,7 @@ declare const SIGNATURE: unique symbol;
 
   ```app/components/cost.gjs
   import formatCurrency from '../helpers/format-currency';
-    
+
   <template>
     <div>{{formatCurrency @cents currency="$"}}</div>
   </template>
@@ -193,10 +193,6 @@ export default class Helper<S = unknown> extends FrameworkObject {
 }
 /* eslint-enable import/export */
 
-export function isClassicHelper(obj: object): boolean {
-  return (obj as any)[IS_CLASSIC_HELPER] === true;
-}
-
 interface ClassicHelperStateBucket {
   instance: HelperInstance<unknown>;
   args: Arguments;
@@ -272,8 +268,6 @@ function isFactoryManager(obj: unknown): obj is InternalFactoryManager<object> {
 setHelperManager((owner: InternalOwner | undefined): ClassicHelperManager => {
   return new ClassicHelperManager(owner);
 }, Helper);
-
-export const CLASSIC_HELPER_MANAGER = getInternalHelperManager(Helper);
 
 ///////////
